feat: makes verify_and_get_contract_params function correcting items order
